plan_mcp_agent/pyproject.toml
[project]
name = "plan-mcp-agent"
version = "0.1.0"
description = "LangGraph agent with planning, OS tools and MCP support - Claude Desktop alternative"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-anthropic>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-mcp-adapters>=0.1.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "asyncio>=3.4.3",
    "aiohttp>=3.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ipython>=8.18.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
